backend/requirements.txt
# Web Framework & Server
fastapi[standard]==0.135.1
uvicorn==0.41.0

# Data & MLB API
python-mlb-statsapi==0.7.2
pandas==3.0.1
pydantic==2.12.5

# Utilities
python-multipart==0.0.20
httpx==0.28.1